refactor: `TraceableExecution` and use standalone execution and trace functions
features:

- rename `TraceableExecution` to `TraceableEngine` to separate engine scope from standalone scope
- refactor `TraceableEngine` to use the common `trace()` function and the `execute()` behind
- doc and generate examples to ensure deprecated engin still work and supported

⚠️ Deprecation Notice:

- `TraceableExecution` is deprecated and will be removed in a future release. Use `TraceableEngine` instead.
