+++ examples/authentication.json
@@ -1,77 +1,77 @@
 [
   {
     "data": {
-      "id": "registerUser_1740075283095_7d0ee34b-4996-41f8-b26f-ff1bc19b4625",
+      "id": "registerUser_1740228152303_b78e0b6a-41a3-4b6e-b7d0-ecfb3d71d56f",
       "label": "registerUser",
       "inputs": [
         "john_doe",
         "secure_password"
       ],
       "outputs": "User john_doe successfully registered",
-      "startTime": "2025-02-20T18:14:43.095Z",
-      "endTime": "2025-02-20T18:14:48.097Z",
-      "duration": 5002.056375026703,
-      "elapsedTime": "5 seconds and 2.056 ms",
+      "startTime": "2025-02-22T12:42:32.303Z",
+      "endTime": "2025-02-22T12:42:37.305Z",
+      "duration": 5002.08141708374,
+      "elapsedTime": "5 seconds and 2.081 ms",
       "narratives": [],
       "parallel": false,
       "abstract": false,
-      "createTime": "2025-02-20T18:14:48.098Z"
+      "createTime": "2025-02-22T12:42:37.306Z"
     },
     "group": "nodes"
   },
   {
     "data": {
-      "id": "loginUser_1740075288098_229fc951-9746-45c1-a05b-2dd96cd92b26",
+      "id": "loginUser_1740228157306_f0bb1cef-6fb6-4ba3-9e17-580450151c47",
       "label": "loginUser",
       "inputs": [
         "john_doe",
         "secure_password"
       ],
       "outputs": "User john_doe successfully logged in",
-      "startTime": "2025-02-20T18:14:48.098Z",
-      "endTime": "2025-02-20T18:14:51.099Z",
-      "duration": 3001.3917499780655,
-      "elapsedTime": "3 seconds and 1.392 ms",
+      "startTime": "2025-02-22T12:42:37.306Z",
+      "endTime": "2025-02-22T12:42:40.308Z",
+      "duration": 3001.641541004181,
+      "elapsedTime": "3 seconds and 1.642 ms",
       "narratives": [],
       "parallel": false,
       "abstract": false,
-      "createTime": "2025-02-20T18:14:51.099Z"
+      "createTime": "2025-02-22T12:42:40.308Z"
     },
     "group": "nodes"
   },
   {
     "data": {
-      "id": "getUserInformation_1740075291100_535d2c3e-9c9f-4f2b-8b42-5a176dcafc63",
+      "id": "getUserInformation_1740228160308_0f23a14d-140d-4642-80b3-869332cab3c5",
       "label": "getUserInformation",
       "inputs": [
         "john_doe"
       ],
       "outputs": "User Information for john_doe: Full Name - John Doe, Email - john.doe@example.com, Role - User",
-      "startTime": "2025-02-20T18:14:51.100Z",
-      "endTime": "2025-02-20T18:14:52.101Z",
-      "duration": 1001.484708070755,
-      "elapsedTime": "1 second and 1.485 ms",
+      "startTime": "2025-02-22T12:42:40.308Z",
+      "endTime": "2025-02-22T12:42:41.309Z",
+      "duration": 1000.6547080278397,
+      "elapsedTime": "1 second and 0.655 ms",
       "narratives": [],
       "parallel": false,
       "abstract": false,
-      "createTime": "2025-02-20T18:14:52.101Z"
+      "createTime": "2025-02-22T12:42:41.309Z"
     },
     "group": "nodes"
   },
   {
     "data": {
-      "id": "registerUser_1740075283095_7d0ee34b-4996-41f8-b26f-ff1bc19b4625->loginUser_1740075288098_229fc951-9746-45c1-a05b-2dd96cd92b26",
-      "source": "registerUser_1740075283095_7d0ee34b-4996-41f8-b26f-ff1bc19b4625",
-      "target": "loginUser_1740075288098_229fc951-9746-45c1-a05b-2dd96cd92b26",
+      "id": "registerUser_1740228152303_b78e0b6a-41a3-4b6e-b7d0-ecfb3d71d56f->loginUser_1740228157306_f0bb1cef-6fb6-4ba3-9e17-580450151c47",
+      "source": "registerUser_1740228152303_b78e0b6a-41a3-4b6e-b7d0-ecfb3d71d56f",
+      "target": "loginUser_1740228157306_f0bb1cef-6fb6-4ba3-9e17-580450151c47",
       "parallel": false
     },
     "group": "edges"
   },
   {
     "data": {
-      "id": "loginUser_1740075288098_229fc951-9746-45c1-a05b-2dd96cd92b26->getUserInformation_1740075291100_535d2c3e-9c9f-4f2b-8b42-5a176dcafc63",
-      "source": "loginUser_1740075288098_229fc951-9746-45c1-a05b-2dd96cd92b26",
-      "target": "getUserInformation_1740075291100_535d2c3e-9c9f-4f2b-8b42-5a176dcafc63",
+      "id": "loginUser_1740228157306_f0bb1cef-6fb6-4ba3-9e17-580450151c47->getUserInformation_1740228160308_0f23a14d-140d-4642-80b3-869332cab3c5",
+      "source": "loginUser_1740228157306_f0bb1cef-6fb6-4ba3-9e17-580450151c47",
+      "target": "getUserInformation_1740228160308_0f23a14d-140d-4642-80b3-869332cab3c5",
       "parallel": false
     },
     "group": "edges"
