+++ examples/greeting.json
@@ -1,7 +1,7 @@
 [
   {
     "data": {
-      "id": "generateGreeting_1740075293454_00ae78e2-1ad5-43d8-b50b-fd1e61d76378",
+      "id": "generateGreeting_1740228162693_80014bc3-1354-4eb2-85a7-51f0df6d471d",
       "label": "1 - generateGreeting",
       "inputs": [
         {
@@ -21,9 +21,9 @@
       "narratives": [
         "here is tracing narrative for greeting John Doe"
       ],
-      "startTime": "2025-02-20T18:14:53.454Z",
+      "startTime": "2025-02-22T12:42:42.693Z",
       "abstract": false,
-      "createTime": "2025-02-20T18:14:53.455Z"
+      "createTime": "2025-02-22T12:42:42.694Z"
     },
     "group": "nodes"
   }
