+++ examples/usage.json
@@ -1,47 +1,47 @@
 [
   {
     "data": {
-      "id": "function_1740075294130_ea6edf05-2fc9-4c8a-bf56-c3eca282fd14",
+      "id": "function_1740228163385_224fdab9-3873-4a03-bfe1-9eb3640af217",
       "label": "function",
       "inputs": [
         "param1"
       ],
       "outputs": "result1 for param1",
-      "startTime": "2025-02-20T18:14:54.130Z",
-      "endTime": "2025-02-20T18:14:54.131Z",
-      "duration": 0.37679195404052734,
-      "elapsedTime": "0.377 ms",
+      "startTime": "2025-02-22T12:42:43.385Z",
+      "endTime": "2025-02-22T12:42:43.386Z",
+      "duration": 0.4349590539932251,
+      "elapsedTime": "0.435 ms",
       "narratives": [],
       "parallel": false,
       "abstract": false,
-      "createTime": "2025-02-20T18:14:54.131Z"
+      "createTime": "2025-02-22T12:42:43.386Z"
     },
     "group": "nodes"
   },
   {
     "data": {
-      "id": "function_1740075294131_a7122250-5441-4cd1-8a74-acedc880a7aa",
+      "id": "function_1740228163386_6ddba430-b906-4082-8f7b-5f36900026a6",
       "label": "function",
       "inputs": [
         "result1 for param1"
       ],
       "outputs": "result2 for result1 for param1",
-      "startTime": "2025-02-20T18:14:54.131Z",
-      "endTime": "2025-02-20T18:14:54.131Z",
-      "duration": 0.2054579257965088,
-      "elapsedTime": "0.205 ms",
+      "startTime": "2025-02-22T12:42:43.386Z",
+      "endTime": "2025-02-22T12:42:43.386Z",
+      "duration": 0.218500018119812,
+      "elapsedTime": "0.219 ms",
       "narratives": [],
       "parallel": false,
       "abstract": false,
-      "createTime": "2025-02-20T18:14:54.131Z"
+      "createTime": "2025-02-22T12:42:43.386Z"
     },
     "group": "nodes"
   },
   {
     "data": {
-      "id": "function_1740075294130_ea6edf05-2fc9-4c8a-bf56-c3eca282fd14->function_1740075294131_a7122250-5441-4cd1-8a74-acedc880a7aa",
-      "source": "function_1740075294130_ea6edf05-2fc9-4c8a-bf56-c3eca282fd14",
-      "target": "function_1740075294131_a7122250-5441-4cd1-8a74-acedc880a7aa",
+      "id": "function_1740228163385_224fdab9-3873-4a03-bfe1-9eb3640af217->function_1740228163386_6ddba430-b906-4082-8f7b-5f36900026a6",
+      "source": "function_1740228163385_224fdab9-3873-4a03-bfe1-9eb3640af217",
+      "target": "function_1740228163386_6ddba430-b906-4082-8f7b-5f36900026a6",
       "parallel": false
     },
     "group": "edges"
